feat: 연락처 추가 기능 구현 및 메세지 출력 로직 변경
