+++ PhoneBookManager/PhoneBookManager/Model/Contact.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-struct Contact {
+struct Contact: Hashable {
     private var name: String
     private var age: Int
     private var phoneNumber: String
