+++ PhoneBookManager/PhoneBookManager/Model/InfoMessage.swift
@@ -1,0 +1,14 @@
+//
+//  InfoMessage.swift
+//  PhoneBookManager
+//
+//  Created by 신동오 on 2022/12/29.
+//
+
+import Foundation
+
+enum InfoMessage {
+    static let menuInput = "1) 연락처 추가 2) 연락처 목록보기 3) 연락처 검색 x) 종료 \n 메뉴를 선택해주세요 : "
+    static let addContact = "연락처 정보를 입력해주세요 : "
+    static let contactAlreadyExists = "이미 존재하는 연락처입니다"
+}
